Handle dynamically created dropdowns better
I'm using dynamically created dropdowns on the  Mortality page, so
selecting sex and country select which sub-page data comes from.
That's a dropdown inside a riskdisplay:
<riskdisplay>Age group: <dropdown table="Mortality/Data/{countrycode}:{sex}" label_column="Age"/></riskdisplay>

That worked nicely, except changing the country or sex (which makes the riskdisplay create a new dropdown pointing to the correct riskdata table) loses the user's selection.

This change fixes that: if the pagestate contains a selection state from a previous dropdown, that pagestate is used to keep that item selected.